rename: NoProfileThumbnailIcon.tsx를 components/profile-images 디렉터리로 이동
