+++ src/components/profile-images/NoProfileThumbnailIcon.tsx
