+++ src/pages/home/components/post/PostItem.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react'
 import { useNavigate, Link } from 'react-router-dom'
 import tw, { styled } from 'twin.macro'
 import { useProfileStore } from '@/stores/userProfileStore'
-import NoProfileThumbnailIcon from '@/pages/search/components/NoProfileThumbnailIcon'
+import NoProfileThumbnailIcon from '@/components/profile-images/NoProfileThumbnailIcon'
 import { getUserInform } from '@/apis/userApis'
 import CommentIcon from '@/assets/icons/comment.svg'
 import HeartIcon from '@/assets/icons/following.svg'
